fix: address iam ownership pr review
