+++ packages/auth-runtime/src/iam-contents/mutation-authorization.ts
@@ -8,7 +8,10 @@ import {
 
 import type { UpdateContentSchemaInput } from './schemas.js';
 import type { ResolvedContentActor } from './request-context.js';
-import { authorizeContentAction, resolveContentAuthorizationPermissions } from './request-context.js';
+import {
+  authorizeContentAction,
+  resolveContentAuthorizationPermissions,
+} from './request-context.js';
 
 const metadataFields = [
   'title',
@@ -68,12 +71,35 @@ export const resolveUpdateContentActions = (
     requiredCapabilities.add('content.update_payload');
   }
 
-  const statusCapability = data.status ? resolveStatusCapability(currentContent.status, data.status) : undefined;
+  const statusCapability = data.status
+    ? resolveStatusCapability(currentContent.status, data.status)
+    : undefined;
   if (statusCapability) {
     requiredCapabilities.add(statusCapability);
   }
 
   return [...requiredCapabilities].map(resolveContentAuthorizationAction);
+};
+
+const resolveProspectiveContentResource = (
+  currentContent: IamContentListItem,
+  data: UpdateContentSchemaInput
+) => ({
+  organizationId: data.organizationId ?? currentContent.organizationId,
+  ownerUserId: data.ownerUserId ?? currentContent.ownerUserId,
+  ownerOrganizationId: data.ownerOrganizationId ?? currentContent.ownerOrganizationId,
+});
+
+const hasDestinationAuthorizationTargetChanged = (
+  currentContent: IamContentListItem,
+  data: UpdateContentSchemaInput
+): boolean => {
+  const nextResource = resolveProspectiveContentResource(currentContent, data);
+  return (
+    nextResource.organizationId !== currentContent.organizationId ||
+    nextResource.ownerUserId !== currentContent.ownerUserId ||
+    nextResource.ownerOrganizationId !== currentContent.ownerOrganizationId
+  );
 };
 
 export const authorizeUpdateContentActions = async (
@@ -83,42 +109,59 @@ export const authorizeUpdateContentActions = async (
   data: UpdateContentSchemaInput
 ): Promise<Response | null> => {
   const actions = resolveUpdateContentActions(currentContent, data);
-  const sourcePermissions = await resolveContentAuthorizationPermissions(actor, currentContent.organizationId);
+  const sourcePermissions = await resolveContentAuthorizationPermissions(
+    actor,
+    currentContent.organizationId
+  );
 
   if ('error' in sourcePermissions) {
     return sourcePermissions.error;
   }
 
   for (const action of actions) {
-    const authorizationError = await authorizeContentAction(actor, action.primitiveAction, {
-      contentId,
-      contentType: currentContent.contentType,
-      domainCapability: action.domainCapability,
-      organizationId: currentContent.organizationId,
-      ownerUserId: currentContent.ownerUserId,
-      ownerOrganizationId: currentContent.ownerOrganizationId,
-    }, { permissions: sourcePermissions.permissions });
+    const authorizationError = await authorizeContentAction(
+      actor,
+      action.primitiveAction,
+      {
+        contentId,
+        contentType: currentContent.contentType,
+        domainCapability: action.domainCapability,
+        organizationId: currentContent.organizationId,
+        ownerUserId: currentContent.ownerUserId,
+        ownerOrganizationId: currentContent.ownerOrganizationId,
+      },
+      { permissions: sourcePermissions.permissions }
+    );
     if (authorizationError) {
       return authorizationError;
     }
   }
 
-  if (data.organizationId && data.organizationId !== currentContent.organizationId) {
-    const destinationPermissions = await resolveContentAuthorizationPermissions(actor, data.organizationId);
+  if (hasDestinationAuthorizationTargetChanged(currentContent, data)) {
+    const nextResource = resolveProspectiveContentResource(currentContent, data);
+    const destinationPermissions = await resolveContentAuthorizationPermissions(
+      actor,
+      nextResource.organizationId
+    );
 
     if ('error' in destinationPermissions) {
       return destinationPermissions.error;
     }
 
     for (const action of actions) {
-      const destinationAuthorizationError = await authorizeContentAction(actor, action.primitiveAction, {
-        contentId,
-        contentType: currentContent.contentType,
-        domainCapability: action.domainCapability,
-        organizationId: data.organizationId,
-        ownerUserId: currentContent.ownerUserId,
-        ownerOrganizationId: data.organizationId,
-      }, { permissions: destinationPermissions.permissions });
+      const destinationAuthorizationError = await authorizeContentAction(
+        actor,
+        action.primitiveAction,
+        {
+          contentId,
+          contentType: currentContent.contentType,
+          domainCapability: action.domainCapability,
+          organizationId: nextResource.organizationId,
+          ownerUserId: nextResource.ownerUserId,
+          ownerOrganizationId: nextResource.ownerOrganizationId,
+        },
+        { permissions: destinationPermissions.permissions }
+      );
       if (destinationAuthorizationError) {
         return destinationAuthorizationError;
       }
